feature: #28: Underline - View Extension
